+++ src/blutwurst/tuple_generator.clj
@@ -37,33 +37,38 @@
 (defn- column-is-string? [column]
   (some #{(:type column)} '("VARCHAR" "NVARCHAR" "CHAR")))
 
-(defn- make-resource-generator-fn [resource garbage-rows]
+(defn- make-resource-generator-fn [resource garbage-rows value-index]
  (let [values (drop garbage-rows (-> resource io/resource io/reader csv/read-csv))
        value-count (count values)]
   (trace values)
   (fn [c]
-   (string/trimr (nth (nth values (random-integer-under-value value-count)) 0)))
+   (string/trimr (nth (nth values (random-integer-under-value value-count)) value-index)))
  ))
 
 (def value-generation-strategies
   ^{ :private true }
   [
+   { 
+     :name "Random State Selector (U.S. and Canada)"
+     :determiner #(and (column-is-string? %) (string/includes? (:name %) "STATE"))
+     :generator (make-resource-generator-fn "state_table.csv" 1 1)
+   }
    {
      :name "Random Full Name Selector (U.S.)"
      :determiner #(and (column-is-string? %) (string/includes? (:name %) "FULL") (string/includes? (:name %) "NAME"))
-     :generator (let [last-name-fn (make-resource-generator-fn "Names_2010Census_Top1000.csv" 3)
-                      first-name-fn (make-resource-generator-fn "census-derived-all-first.csv" 0)]
+     :generator (let [last-name-fn (make-resource-generator-fn "Names_2010Census_Top1000.csv" 3 0)
+                      first-name-fn (make-resource-generator-fn "census-derived-all-first.csv" 0 0)]
                   #(str (last-name-fn %) " " (first-name-fn %)))
    }
    {
      :name "Random First Name Selector (U.S.)"
      :determiner #(and (column-is-string? %) (string/includes? (:name %) "FIRST"))
-     :generator (make-resource-generator-fn "census-derived-all-first.csv" 0)
+     :generator (make-resource-generator-fn "census-derived-all-first.csv" 0 0)
    }
    {
       :name "Random Last Name Selector (U.S.)"
       :determiner #(and (column-is-string? %) (string/includes? (:name %) "LAST"))
-      :generator (make-resource-generator-fn "Names_2010Census_Top1000.csv" 3)
+      :generator (make-resource-generator-fn "Names_2010Census_Top1000.csv" 3 0)
    }
    {
       :name "Random String Generator"
@@ -87,19 +92,31 @@
    }
   ])
 
-(defn- select-generators-for-column [column]
-  (let [generator-search (fn [column strategies]
-                           (if (empty? strategies)
-                             nil
-                             (let [current-strategy (first strategies)]
-                               (if ((:determiner current-strategy) column)
-                                 current-strategy
-                                 (recur column (rest strategies))
-                                 ))
-                             ))]
+(defn- generator-search [column strategies]
+   (if (empty? strategies)
+     nil
+     (let [current-strategy (first strategies)]
+       (if ((:determiner current-strategy) column)
+         current-strategy
+         (recur column (rest strategies))
+         ))
+     ))
 
-    (generator-search column value-generation-strategies)
-    ))
+(defn- find-override-for-column [spec column]
+ (:generator-name (first (filter #(re-matches (:column-pattern %) (:name column)) (:column-generator-overrides spec)))))
+
+(defn- find-generator-by-name [generator-name strategies]
+ (first (filter #(= (:name %) generator-name) strategies)))
+
+(defn- select-generators-for-column [spec column]
+  (let [override-name (find-override-for-column spec column)]
+
+    (if (nil? override-name)
+      (generator-search column value-generation-strategies)
+      (do 
+        (trace "Found override " override-name " for column " column)
+        (find-generator-by-name override-name value-generation-strategies))
+    )))
 
 (defn- build-dependency-selector-fn [generated-data dependency]
   (let [targeted-table (first (filter #(and (= (-> % :table :schema) (:target-schema dependency))
@@ -118,9 +135,10 @@
   (remove #(some (fn [x] (= (:name %) x)) linked-columns) (:columns table))
  ))
 
-(defn- build-generator-fn [table generated-data]
+(defn- build-generator-fn [spec table generated-data]
  (let [dependency-selectors (map (partial build-dependency-selector-fn generated-data) (:dependencies table))
-       value-generators (map #(hash-map :column % :generator (-> % select-generators-for-column :generator)) 
+       value-generators (map #(hash-map :column % 
+                                        :generator (->> % (select-generators-for-column spec) :generator)) 
                              (value-columns table))]
 
     (doseq [c value-generators]
@@ -137,24 +155,24 @@
     )
   ))
 
-(defn- generate-tuples-for-table [table-descriptor number-of-tuples generated-data]
-  (let [table-generator (build-generator-fn table-descriptor generated-data)]
-    {:table table-descriptor :tuples (repeatedly number-of-tuples table-generator)}
+(defn- generate-tuples-for-table [table-descriptor spec generated-data]
+  (let [table-generator (build-generator-fn spec table-descriptor generated-data)]
+    {:table table-descriptor :tuples (repeatedly (:number-of-rows spec) table-generator)}
     ))
 
-(defn- generate-tuples-for-plan* [rows-to-generate execution-plan result]
+(defn- generate-tuples-for-plan* [spec execution-plan result]
  (if (empty? execution-plan)
   (reverse result)
   (do
       (trace "Beginning data generation for table " (first execution-plan))
-      (recur rows-to-generate
+      (recur spec
              (rest execution-plan)
-             (cons (generate-tuples-for-table (first execution-plan) rows-to-generate result) result)))
+             (cons (generate-tuples-for-table (first execution-plan) spec result) result)))
  ))
 
 (defn generate-tuples-for-plan [spec execution-plan]
  (trace "Beginning tuple generation for execution plan.")
- (generate-tuples-for-plan* (:number-of-rows spec) execution-plan '()))
+ (generate-tuples-for-plan* spec execution-plan '()))
  
 (defn retrieve-registered-generators [] 
  (map :name value-generation-strategies))