+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to this project will be documented in this file. This change
 - Added composite generator for full names.
 - Added `--list-generators` command-line option so that the user can see what generators are provided.
 - Added command line option (`-s`) to limit the schemas being scanned. By default, all schemas are scanned but provision of a list of schemas will limit scans to those.
+- Added pair of command line options (`--column` and `--generator`) to allow a user to override the matched generators.
 
 ## [0.2.0] - 2017-08-17
 ### Added